refactor(llm-agent): move ClineClientAdapter from llm-agent-server
Pure file relocation. ClineClientAdapter is now exported from llm-agent,
with a re-export shim in the server for backward compatibility.
